Run JS data clearing alongside native storage clearing
Fire button cleanup with the duckAiNativeStorage flag on previously left
IndexedDB and localStorage traces because the JS layer is kept in sync
even when native storage is the system of record. HistoryCleaner now
runs the c-s-s duck-ai-data-clearing path alongside the native clear
via an injectable AIChatJSDataCleaning.
